[NXP][platform][common] Fix build warning
(cherry picked from commit 72b1cbaaf36fe6cf2af57a8c7ad3cc457b45c713)

+++ examples/platform/nxp/common/app_task/source/AppTaskBase.cpp
@@ -436,5 +436,5 @@ void chip::NXP::App::AppTaskBase::PrintCurrentVersion()
     err = ConfigurationMgr().GetSoftwareVersion(currentVersion);
     ReturnOnFailure(err);
 
-    ChipLogProgress(DeviceLayer, "Current Software Version: %s, %lu", currentSoftwareVer, currentVersion);
+    ChipLogProgress(DeviceLayer, "Current Software Version: %s, %d", currentSoftwareVer, static_cast<int>(currentVersion));
 }
